Fix a test now that implicit client::Object to double cast is not allowed

+++ tests/unit/dom/test2.cpp
@@ -9,6 +9,6 @@ void webMain()
 {
 	//Test Object to Number conversion
 	client::Object* o = client::JSON.parse("42.42");
-	double dval = *o;
+	double dval = (double)*o;
 	assertEqual(dval, 42.42, "Object to Number");
 }
